Update caesar_cipher.py
Added more logics like asking the user if they want to continue 'encoding' and 'decoding'. Also applied logic to cipher to check if the user enters letters to be encrypted or decrypted. If letters, encrypt or decrypt them. If number or symbols are included, only encrypt or decrypt the letters and keep the numbers or symbols as is. The shift position was also modified to ensure that whatever number the user enters, it doesn't go out of the 'alphabets' list range. For example, if the user enters 86, the shift will return the reminder of the shift value divided by the number of items in the 'alphabets' list.