Update snake.py
In snake game, snake cannot move back on itself. I added a if statement to each keystroke function to stop snake game from moving back on itself. For example, previously, before adding the  the "if" statements, when the snake is facing down, when you press up the snake head changes to up. With the if statements, when the snake is facing up or down or any direction, when you press the opposite direction, the snake head does not change to that direction.